Moved context menus to a separate component
