+++ frontend/src/components/GraphView/GraphView.js
@@ -1,48 +1,13 @@
-import { useEffect, useRef, useCallback, useState } from 'react'
+import { useEffect, useRef, useCallback } from 'react'
 
-import { GraphContent, SelectionBox, Dropdown } from '/src/components'
+import { GraphContent, SelectionBox } from '/src/components'
 import { MarkerProvider } from '/src/providers'
 import { useViewStore } from '/src/stores'
 import { VIEW_MOVE_STEP, GRID_SNAP } from '/src/config/interactions'
 
 import { Svg } from './graphViewStyle'
+import { ContextMenus } from './components'
 import { useViewDragging } from './hooks'
-
-const stateContextItems = [
-  {
-    label: 'Set as initial',
-    action: 'SET_STATE_INITIAL',
-  },
-  {
-    label: 'Set as final',
-    action: 'SET_STATE_FINAL',
-  },
-  'hr',
-  {
-    label: 'Set  label',
-    action: 'SET_STATE_LABEL',
-  },
-  'hr',
-  {
-    label: 'Change name',
-    action: 'SET_STATE_NAME',
-  },
-  'hr',
-  {
-    label: 'Copy',
-    action: 'COPY',
-  },
-  {
-    label: 'Paste',
-    action: 'PASTE',
-  },
-  'hr',
-  {
-    label: 'Delete',
-    shortcut: '⌫',
-    action: 'DELETE'
-  },
-]
 
 const GraphView = props => {
   const containerRef = useRef()
@@ -82,17 +47,6 @@ const GraphView = props => {
     return () => window.removeEventListener('keydown', onKeyDown)
   }, [])
 
-  // TODO: Move this and the context menu to a separate file
-  const [stateContext, setStateContext] = useState({ visible: false })
-  const onStateContext = useCallback(({ detail: { states, x, y } }) => {
-    setStateContext({ visible: true, x, y })
-  }, [])
-
-  useEffect(() => {
-    document.addEventListener('stateContext', onStateContext)
-    return () => document.removeEventListener('stateContext', onStateContext)
-  }, [])
-
   // Determine svg background (grid)
   const backgroundPosition = `${-position.x / scale}px ${-position.y / scale}px`
   const backgroundSize = `${1 / scale * GRID_SNAP * 2}px ${1 / scale * GRID_SNAP * 2}px`
@@ -119,15 +73,7 @@ const GraphView = props => {
         </MarkerProvider>
       </Svg>
 
-      <Dropdown
-        visible={stateContext.visible}
-        onClose={() => setStateContext({ visible: false })}
-        style={{
-          top: `${stateContext.y}px`,
-          left: `${stateContext.x}px`,
-        }}
-        items={stateContextItems}
-      />
+      <ContextMenus />
     </>
   )
 }
