+++ frontend/src/components/GraphView/components/index.js
@@ -1,0 +1,1 @@
+export { default as ContextMenus } from './ContextMenus/ContextMenus'
